+++ CHANGELOG.rst
@@ -17,6 +17,34 @@ Unreleased
 ----------
 * nothing unreleased
 
+[5.13.3]
+--------
+* fix: fixed chat_completion endpoint to handle Xpert AI v2 response correctly.
+
+[5.13.2]
+--------
+* fix: create migration for unmanage django model that points at DB view.
+
+[5.13.1]
+--------
+* fix: enterprise-customer-support sorting/pagination using view
+
+[5.13.0]
+--------
+* feat: adds EnterpriseCustomerAdmin model
+
+[5.12.10]
+---------
+* fix: fixes datetime attribute error
+
+[5.12.9]
+--------
+* fix: sync ecu table user_fk changes with django
+
+[5.12.8]
+--------
+* fix: fixing enrollments bug in members endpoint
+
 [5.12.7]
 --------
 * feat: adding OnboardingFlow model